auto/data-collector/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-data-collector"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "pyarrow",
    "numpy",
    "requests",
    "questdb",
    "python-dateutil",
    "prometheus_client>=0.20",
]

[project.optional-dependencies]
dev = ["pytest", "python-dotenv"]

[project.scripts]
crypto-data-collector = "crypto_data_collector.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
